actually fix headers support on node:http and proper error on non-ASCII chars on fetch
