+++ src/bun.js/bindings/webcore/HTTPParsers.cpp
@@ -125,10 +125,14 @@ bool isValidHTTPHeaderValue(const String& value)
     c = value[value.length() - 1];
     if (isTabOrSpace(c))
         return false;
+
     if (value.is8Bit()) {
         const LChar* begin = value.span8().data();
         const LChar* end = begin + value.length();
         for (const LChar* p = begin; p != end; ++p) {
+            if (!isASCII(*p)) {
+                return false;
+            }
             if (UNLIKELY(*p <= 13)) {
                 LChar c = *p;
                 if (c == 0x00 || c == 0x0A || c == 0x0D)
@@ -138,6 +142,9 @@ bool isValidHTTPHeaderValue(const String& value)
     } else {
         for (unsigned i = 0; i < value.length(); ++i) {
             c = value[i];
+            if (!isASCII(c)) {
+                return false;
+            }
             if (c == 0x00 || c == 0x0A || c == 0x0D || c > 0x7F)
                 return false;
         }
